+++ src/test/java/io/github/astrapi69/swing/panels/tree/TestTreeElementPanel.java
@@ -24,6 +24,7 @@
  */
 package io.github.astrapi69.swing.panels.tree;
 
+import java.awt.*;
 import java.awt.event.MouseEvent;
 
 import javax.swing.*;
@@ -35,14 +36,20 @@ import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.DefaultTreeModel;
 import javax.swing.tree.TreeModel;
 import javax.swing.tree.TreePath;
+import javax.swing.tree.TreeSelectionModel;
+
+import io.github.astrapi69.swing.tree.TreeNodeFactory;
+import org.jdesktop.swingx.JXTree;
 
 import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.Model;
+import io.github.astrapi69.swing.dialog.JOptionPaneExtensions;
+import io.github.astrapi69.swing.listener.RequestFocusListener;
+import io.github.astrapi69.swing.tree.renderer.TreeNodeCellRenderer;
 import io.github.astrapi69.tree.TreeElement;
 import io.github.astrapi69.tree.TreeNode;
-import io.github.astrapi69.tree.api.ITreeNode;
 
-public class TestTreeElementPanel extends JTreePanel<TreeNode<TreeElement>>
+public class TestTreeElementPanel extends TreeElementPanel
 {
 
 	private static final long serialVersionUID = 1L;
@@ -58,9 +65,9 @@ public class TestTreeElementPanel extends JTreePanel<TreeNode<TreeElement>>
 	}
 
 	@Override
-	protected JTree newTree()
+	protected JXTree newTree()
 	{
-		JTree tree = super.newTree();
+		JXTree tree = super.newTree();
 		tree.setCellRenderer(new TreeNodeCellRenderer<TreeElement>());
 		return tree;
 	}
@@ -69,8 +76,13 @@ public class TestTreeElementPanel extends JTreePanel<TreeNode<TreeElement>>
 	protected TreeModel newTreeModel(final Model<TreeNode<TreeElement>> model)
 	{
 		TreeNode<TreeElement> parentTreeNode = model.getObject();
+		TreeModel treeModel;
 
-		TreeModel treeModel = new TreeNodeModel(parentTreeNode);
+		// treeModel = new TreeNodeModel(parentTreeNode);
+
+		DefaultMutableTreeNode rootNode = TreeNodeFactory.newDefaultMutableTreeNode(parentTreeNode);
+
+		treeModel = new DefaultTreeModel(rootNode, true);
 
 		treeModel.addTreeModelListener(new TreeModelListener()
 		{
@@ -82,6 +94,8 @@ public class TestTreeElementPanel extends JTreePanel<TreeNode<TreeElement>>
 				node = (DefaultMutableTreeNode)lastPathComponent;
 				int index = e.getChildIndices()[0];
 				node = (DefaultMutableTreeNode)(node.getChildAt(index));
+
+
 			}
 
 			@Override
@@ -102,37 +116,13 @@ public class TestTreeElementPanel extends JTreePanel<TreeNode<TreeElement>>
 		return treeModel;
 	}
 
-	protected void onInitializeGroupLayout()
-	{
-		javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
-		this.setLayout(layout);
-		layout.setHorizontalGroup(
-			layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-				.addGroup(layout.createSequentialGroup().addContainerGap()
-					.addComponent(scrTree, javax.swing.GroupLayout.PREFERRED_SIZE, 384,
-						javax.swing.GroupLayout.PREFERRED_SIZE)
-					.addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)));
-		layout.setVerticalGroup(layout
-			.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-			.addGroup(layout.createSequentialGroup().addContainerGap()
-				.addComponent(scrTree, javax.swing.GroupLayout.DEFAULT_SIZE, 536, Short.MAX_VALUE)
-				.addContainerGap()));
-	}
-
-	@Override
-	protected void onInitializeLayout()
-	{
-		super.onInitializeLayout();
-		onInitializeGroupLayout();
-	}
-
 	@Override
 	protected void onSingleClick(MouseEvent e)
 	{
 		int x = e.getX();
 		int y = e.getY();
 		TreePath selectionPath = tree.getPathForLocation(e.getX(), e.getY());
-
+		TreeSelectionModel selectionModel = tree.getSelectionModel();
 		JPopupMenu popup = new JPopupMenu();
 		JMenuItem addChild = new JMenuItem("add node...");
 		addChild.addActionListener(le -> {
@@ -157,23 +147,46 @@ public class TestTreeElementPanel extends JTreePanel<TreeNode<TreeElement>>
 					}
 				}
 			});
-			Object[] inputFields = { "Enter name for node", textField1, "Is leaf", checkBox };
+			// Object[] inputFields = { "Enter name for node", textField1, "Is leaf", checkBox };
+			//
+			// int option = JOptionPane.showConfirmDialog(this, inputFields, "Multiple Inputs",
+			// JOptionPane.OK_CANCEL_OPTION, JOptionPane.INFORMATION_MESSAGE);
 
-			int option = JOptionPane.showConfirmDialog(this, inputFields, "Multiple Inputs",
-				JOptionPane.OK_CANCEL_OPTION, JOptionPane.INFORMATION_MESSAGE);
+			JPanel panel = new JPanel(new GridLayout(2, 2));
+			panel.add(new JLabel("Enter name for node:"));
+			panel.add(textField1);
+			panel.add(new JLabel("Is leaf:"));
+			panel.add(checkBox);
+
+			JOptionPane pane = new JOptionPane(panel, JOptionPane.INFORMATION_MESSAGE,
+				JOptionPane.OK_CANCEL_OPTION);
+			JDialog dialog = pane.createDialog(null, "New node");
+			dialog.addWindowFocusListener(new RequestFocusListener(textField1));
+			dialog.pack();
+			dialog.setLocationRelativeTo(null);
+			dialog.setVisible(true);
+			int option = JOptionPaneExtensions.getSelectedOption(pane);
 
 			if (option == JOptionPane.OK_OPTION)
 			{
 				Object lastPathComponent = selectionPath.getLastPathComponent();
-				TreeNode<TreeElement> selectedTreeNode = (TreeNode<TreeElement>)lastPathComponent;
+				DefaultMutableTreeNode selectedTreeNode = (DefaultMutableTreeNode)lastPathComponent;
 
 				boolean allowsChildren = !checkBox.isSelected();
 				String userObject = textField1.getText();
-
+				TreeNode<TreeElement> parentTreeNode = (TreeNode<TreeElement>)selectedTreeNode
+					.getUserObject();
 				TreeElement treeElement = TreeElement.builder().name(userObject)
+					.parent(parentTreeNode.getValue())
 					.node(allowsChildren).build();
-				TreeNode<TreeElement> newEl = new TreeNode<>(treeElement);
-				selectedTreeNode.addChild(newEl);
+				TreeNode<TreeElement> newTreeNode = TreeNode.<TreeElement> builder()
+					.value(treeElement).parent(parentTreeNode).displayValue(userObject)
+					.node(allowsChildren).build();
+				new TreeNode<>(treeElement);
+				DefaultMutableTreeNode newChild = new DefaultMutableTreeNode(newTreeNode,
+					allowsChildren);
+				selectedTreeNode.add(newChild);
+				((DefaultTreeModel)tree.getModel()).reload(selectedTreeNode);
 				tree.treeDidChange();
 			}
 
@@ -183,17 +196,18 @@ public class TestTreeElementPanel extends JTreePanel<TreeNode<TreeElement>>
 		JMenuItem deleteNode = new JMenuItem("delete");
 		deleteNode.addActionListener(le -> {
 
-			Object lastPathComponent = selectionPath.getLastPathComponent();
-			TreeNode<TreeElement> selectedTreeNode = (TreeNode<TreeElement>)lastPathComponent;
-			if (!selectedTreeNode.isRoot())
-			{
-				ITreeNode<TreeElement> parent = selectedTreeNode.getParent();
-				boolean remove = parent.getChildren().remove(selectedTreeNode);
-				System.out.println(remove);
-				TreeNode<TreeElement> root = (TreeNode<TreeElement>)parent.getRoot();
-				TreeNodeModel<TreeElement> ntnm = new TreeNodeModel<TreeElement>(root);
-				getTree().setModel(ntnm);
+			DefaultMutableTreeNode selectedPathComponent = (DefaultMutableTreeNode)selectionPath
+				.getLastPathComponent();
 
+			if (!selectedPathComponent.isRoot())
+			{
+				int selectedNodeIndex = selectedPathComponent.getParent()
+					.getIndex(selectedPathComponent);
+				selectedPathComponent.removeAllChildren();
+				((DefaultMutableTreeNode)selectedPathComponent.getParent())
+					.remove(selectedNodeIndex);
+				((DefaultTreeModel)tree.getModel()).reload(selectedPathComponent);
+				tree.treeDidChange();
 			}
 			else
 			{